Making badges modal responsive for mobile devices
